New Playground Demo

+++ Assets/_Procedural Room/Scripts/Scriptables/SO_Decorations.cs
@@ -8,7 +8,7 @@ public class SO_Decorations : ScriptableObject
     #region Variables / Components
 
     [SerializeField] private GameObject _prefab = default;
-    [SerializeField] private int _size = 0;
+    [SerializeField] private float _size = 0;
     [SerializeField] private Vector3 _positionOffSet = Vector3.zero;
     [SerializeField] private Vector3 _rotationOffSet = Vector3.zero;
     [SerializeField] private bool _allowRandomRotation = false;
@@ -18,7 +18,7 @@ public class SO_Decorations : ScriptableObject
     #region Properties
 
     public GameObject prefab => _prefab;
-    public int size => _size;
+    public float size => _size;
     public Vector3 positionOffSet => _positionOffSet;
     public Vector3 rotationOffSet => _rotationOffSet;
     public bool allowRandomRotation => _allowRandomRotation;
